Moved all default frozen hearts render to a more accessible class, in order to access it from both HealthOverlay compat mixins, and avoid duplicates
